requirements.txt
streamlit
Pillow
IPython
numpy
scipy
matplotlib
weasyprint
biopython
scikit-learn
RNA